server/requirements.txt
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
